@86epxqyup #Renaming textfield and reimporting jest tools from common
