+++ packages/core-library/components/index.ts
@@ -53,3 +53,4 @@ export * from "./Checkbox/Checkbox";
 export * from "./RichTextEditor/RichTextEditor";
 export * from "./blocks/AnswerOptions/AnswerOptions";
 export * from "./Radio/Radio"
+export * from "./Textfield/TextField"
